requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pymssql==2.2.8
email-validator==2.0.0
fastapi-pagination==0.12.5
# Alternative SQL Server drivers (uncomment if needed):
# pyodbc==5.0.1
# Optional PostgreSQL support (install if needed):
# psycopg2-binary==2.9.9
